[Player] integrated penguin model to the game
