--- Assets/Materials/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ed265305a2a0f9d41b3b11f288d48bc4
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
