--- Assets/Materials/Player/SkiGear.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: edf2e28fe724cdc43a85daa12aeb80f3
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
