--- Assets/Materials/Player/SkiSuit.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 69a0864a3145cf848b5f7dfd295cdb8f
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
