+++ Assets/Materials/Players/Penguin.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 53e9a84cbb735fc46aeb8d0df61f81de
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
