+++ Assets/Materials/Players/Penguin/beakAndFeet.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b0fd317153575584c86908158b9f2839
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
