+++ Assets/Materials/Players/Penguin/eyes.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cb24027137638e649b498cf5078a7b7c
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
