+++ Assets/Materials/Players/Penguin/penguinSkin.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fa6988d827736354fa70760718b8ce38
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
