+++ Assets/Materials/Players/Penguin/redTissue.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5c028d03c38d828408d0ec74fdffdf89
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
