+++ Assets/Materials/Players/Penguin/skiGear.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 167fe6a0f605359498422c61610a2e72
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
