+++ Assets/Materials/Players/Penguin/tongue.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9b25559e488482f4ab697fe17fba52a9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
