+++ Assets/Materials/Players/Penguin/whiteFur.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0106337fdc5981f4a8761657de4c9ee4
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
