+++ Assets/Materials/Terrain/SnowyTerrain.mat
@@ -143,7 +143,7 @@ Material:
     - _ReconstructNormals: 1
     - _Smoothness: 0.5
     - _SmoothnessTextureChannel: 0
-    - _SnowBaseDepth: 40
+    - _SnowBaseDepth: 60
     - _SnowDepthCm: 35
     - _SpecularHighlights: 1
     - _SrcBlend: 1
