--- Assets/Prefabs/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: e0f420e7bc91e2842a65b8a10e67dda1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
