--- Assets/Prefabs/Player/Player.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 26cb84ca751b9e44da46bdfd804f6045
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
