--- Assets/Scripts/Player.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 191bbc3a9dce2874ebcd11c36b832596
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
