--- Assets/Scripts/Player/Respawner.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d535aea7ca75e614788d74f0f5052f90
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
