+++ Assets/Scripts/Players/PhysicsSkierController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 32dd580e03240b040b55783fb91204e8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
