+++ Assets/Scripts/Players/SpawnPointManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d1efe89b5a84f7846ba4e79c7ce5c243
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
